moved metrics
simplified business logic
