+++ src/app/metrics/__init__.py
@@ -1,2 +1,2 @@
 # Import metric implementations to trigger registration
-from . import implementations
+from . import bus_factor, code_quality, dataset_and_code, dataset_quality, license_metric, performance_claims, ramp_up_time, size
