--- src/app/metrics/implementations/__init__.py
@@ -1,9 +1,0 @@
-# Import all metric implementations to trigger registration
-from . import license
-from . import bus_factor  
-from . import ramp_up_time
-from . import performance_claims
-from . import size
-from . import dataset_and_code
-from . import code_quality
-from . import dataset_quality
