--- src/app/metrics/implementations/bus_factor.py
@@ -1,71 +1,0 @@
-"""
-Maintainer health and bus factor metric.
-
-Evaluates project sustainability by analyzing contributor diversity,
-activity patterns, and organizational backing.
-"""
-
-from ..base import ResourceBundle
-from ..registry import register
-from ..base_metric import BaseMetric, stable_01
-
-
-@register("bus_factor")
-class BusFactorMetric(BaseMetric):
-    """
-    Evaluates contributor redundancy and maintainer health.
-    
-    Bus factor measures project sustainability risk by assessing:
-    - Number of active contributors and their contribution distribution
-    - Recent commit activity patterns and consistency
-    - Maintainer responsiveness to issues and pull requests
-    - Organizational vs individual project ownership
-    - Community engagement and external contributions
-    
-    Future Implementation Notes:
-    - Query GitHub API for comprehensive contributor statistics
-    - Analyze commit patterns, frequency, and author diversity over time
-    - Check issue response times, PR merge rates, and community engagement
-    - Evaluate organizational backing vs individual maintainer dependency
-    - Consider Hugging Face organization memberships and verified accounts
-    - Assess project governance and decision-making structure
-    
-    Scoring Rubric (Future):
-    - 1.0: Multiple active maintainers, strong organizational backing
-    - 0.8: Several contributors, good activity, some organizational support
-    - 0.6: Few active contributors but consistent maintenance
-    - 0.4: Limited contributors, irregular maintenance activity
-    - 0.2: Single maintainer or very infrequent updates
-    - 0.0: Abandoned project or no visible maintenance
-    """
-    
-    name = "bus_factor"
-    
-    def _compute_score(self, resource: ResourceBundle) -> float:
-        """
-        Evaluate project maintainer health and contributor diversity.
-        
-        TODO: Replace with real implementation that:
-        1. Fetches contributor data from GitHub API (commits, PRs, issues)
-        2. Analyzes commit frequency and author distribution patterns
-        3. Checks recent activity patterns (commits, releases, issue responses)
-        4. Evaluates issue/PR response times and community engagement
-        5. Assesses organizational backing and verified maintainer status
-        6. Measures project governance maturity and decision-making process
-        
-        Args:
-            resource: Bundle containing repository URLs to analyze
-            
-        Returns:
-            Score from 0-1 where 1 = healthy contributor ecosystem with low bus factor risk
-        """
-        # PLACEHOLDER: Real implementation will analyze GitHub contributor
-        # statistics, commit patterns, and organizational backing
-        return stable_01(resource.model_url, "bus_factor")
-    
-    def _get_computation_notes(self, resource: ResourceBundle) -> str:
-        return (
-            f"Placeholder contributor analysis for {resource.model_url}. "
-            "Real implementation will analyze GitHub contributor patterns, "
-            "activity metrics, and organizational backing via GitHub API."
-        )
