--- src/app/metrics/implementations/performance_claims.py
@@ -1,72 +1,0 @@
-"""
-Performance claims and benchmark evidence metric.
-
-Evaluates the credibility and evidence supporting performance claims
-for models through benchmark results and empirical validation.
-"""
-
-from ..base import ResourceBundle
-from ..registry import register
-from ..base_metric import BaseMetric, stable_01
-
-
-@register("performance_claims")
-class PerformanceClaimsMetric(BaseMetric):
-    """
-    Evaluates evidence supporting model performance claims.
-    
-    This metric assesses the credibility of performance assertions by analyzing:
-    - Presence of benchmark results and evaluation metrics
-    - Quality and comprehensiveness of evaluation datasets
-    - Reproducibility of reported results
-    - Comparison with baseline models and state-of-the-art
-    - Methodological rigor in performance evaluation
-    - Independent validation and peer review evidence
-    
-    Future Implementation Notes:
-    - Parse model cards for benchmark results and evaluation metrics
-    - Cross-reference claims with papers, leaderboards, and external evaluations
-    - Analyze evaluation dataset quality and relevance
-    - Check for reproducibility artifacts (code, data, configs)
-    - Validate benchmark methodologies against best practices
-    - Search for independent evaluations and peer review
-    
-    Scoring Rubric (Future):
-    - 1.0: Comprehensive benchmarks with reproducible results and peer validation
-    - 0.8: Good benchmark coverage with some reproducibility evidence
-    - 0.6: Basic performance metrics with limited validation
-    - 0.4: Minimal benchmark results or unclear methodology
-    - 0.2: Vague performance claims without supporting evidence
-    - 0.0: No performance information or unsubstantiated claims
-    """
-    
-    name = "performance_claims"
-    
-    def _compute_score(self, resource: ResourceBundle) -> float:
-        """
-        Evaluate evidence supporting model performance claims.
-        
-        TODO: Replace with real implementation that:
-        1. Parse Hugging Face model cards for performance metrics and benchmarks
-        2. Cross-reference with academic papers and technical reports
-        3. Search for independent evaluations and leaderboard entries
-        4. Analyze benchmark dataset quality and evaluation methodology
-        5. Check for reproducibility artifacts and validation evidence
-        6. Assess performance claims against known baselines and SOTA
-        
-        Args:
-            resource: Bundle containing model and evaluation URLs
-            
-        Returns:
-            Score from 0-1 where 1 = well-evidenced performance claims with validation
-        """
-        # PLACEHOLDER: Real implementation will analyze benchmark results,
-        # model cards, and cross-reference with external validation sources
-        return stable_01(resource.model_url, "performance_claims")
-    
-    def _get_computation_notes(self, resource: ResourceBundle) -> str:
-        return (
-            f"Placeholder performance analysis for {resource.model_url}. "
-            "Real implementation will evaluate benchmark evidence, model cards, "
-            "and cross-reference with papers and leaderboards."
-        )
