--- src/app/metrics/implementations/ramp_up_time.py
@@ -1,73 +1,0 @@
-"""
-Documentation quality and ramp-up time metric.
-
-Evaluates how quickly developers can start using a model by analyzing
-documentation completeness, clarity, and the presence of working examples.
-"""
-
-from ..base import ResourceBundle
-from ..registry import register
-from ..base_metric import BaseMetric, stable_01
-
-
-@register("ramp_up_time")
-class RampUpTimeMetric(BaseMetric):
-    """
-    Evaluates ease of adoption from documentation quality.
-    
-    This metric assesses how quickly developers can start using a model by evaluating:
-    - README completeness and clarity
-    - Usage examples and code snippets  
-    - API documentation quality
-    - Model card information completeness
-    - Quick-start guides and tutorials
-    - Installation/setup instructions
-    
-    Future Implementation Notes:
-    - Fetch and analyze README.md structure and content depth
-    - Parse and validate code examples for syntax correctness
-    - Evaluate Hugging Face model card completeness score
-    - Assess documentation readability using Flesch-Kincaid metrics
-    - Check for step-by-step tutorials and getting-started guides
-    - Verify links and references are working and up-to-date
-    
-    Scoring Rubric (Future):
-    - 1.0: Comprehensive docs with working examples and clear tutorials
-    - 0.8: Good documentation with minor gaps in examples or setup
-    - 0.6: Basic documentation present but missing key information
-    - 0.4: Minimal documentation, unclear setup instructions
-    - 0.2: Very limited documentation, difficult to get started
-    - 0.0: No meaningful documentation or examples
-    """
-    
-    name = "ramp_up_time"
-    
-    def _compute_score(self, resource: ResourceBundle) -> float:
-        """
-        Evaluate documentation quality and developer onboarding ease.
-        
-        TODO: Replace with real implementation that:
-        1. Fetches README and documentation files via GitHub API
-        2. Analyzes content structure, completeness, and readability
-        3. Validates code examples for syntax correctness and executability
-        4. Checks for quick-start sections, tutorials, and setup guides
-        5. Evaluates Hugging Face model card information depth and clarity
-        6. Assesses link validity and reference accuracy
-        7. Measures documentation freshness and maintenance
-        
-        Args:
-            resource: Bundle containing model and documentation URLs
-            
-        Returns:
-            Score from 0-1 where 1 = excellent documentation enabling rapid adoption
-        """
-        # PLACEHOLDER: Real implementation will analyze README quality,
-        # code examples, and model card completeness
-        return stable_01(resource.model_url, "ramp_up")
-    
-    def _get_computation_notes(self, resource: ResourceBundle) -> str:
-        return (
-            f"Placeholder documentation analysis for {resource.model_url}. "
-            "Real implementation will evaluate README quality, code examples, "
-            "and model card completeness via GitHub and HuggingFace APIs."
-        )
